% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrencyConverter.R
\name{DollarToRupee}
\alias{DollarToRupee}
\title{Provides gallon to liters conversion}
\usage{
DollarToRupee(dol)
}
\arguments{
\item{dol}{A value parameter.}
}
\value{
rupee A value.
}
\description{
Provides gallon to liters conversion
}
